# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple api for storing files.",
        "title": "Simple Store",
        "contact": {
            "email": "info@karimi.dev"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/{dir}/{filename}": {
            "get": {
                "description": "Upload file",
                "summary": "Download file",
                "operationId": "file.download",
                "parameters": [
                    {
                        "type": "string",
                        "description": "directory for file",
                        "name": "dir",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name for file",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Upload file",
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Upload file",
                "operationId": "file.upload",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file to upload",
                        "name": "data",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "directory for file",
                        "name": "dir",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name for file",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}